Copy route MTU from net namespace and adjust for network header size
In the original version of this commit
https://github.com/google/gvisor/pull/11857, If r.mtu has been set we skip
calculating the proper MTU in tcpip/stack/route.go:Route.MTU() via endpoint.MTU
and use r.mtu instead. However nothing in the callstack from
tcpip/stack/stack.go:FindRoute to setting r.mtu actually adjusts the size to be
able to hold the endpoint header. Hence we need to adjust it.

Fixes #11856

PiperOrigin-RevId: 843416148